+++ src/api-legacy/v1/borehole/geojson.py
@@ -14,6 +14,8 @@ class ListGeojson(Action):
 
         chronostratigraphy_where, chronostratigraphy_params, chronostratigraphy_joins = self.filterChronostratigraphy(filter)
 
+        lithostratigraphy_where, lithostratigraphy_params, lithostratigraphy_joins = self.filterLithostratigraphy(filter)
+
         casing_where, casing_params, casing_joins = self.filterCasings(filter)
 
         backfill_where, backfill_params, backfill_joins = self.filterBackfill(
@@ -96,6 +98,41 @@ class ListGeojson(Action):
                 ) as chronostratigraphy
                 ON 
                     borehole.id_bho = chronostratigraphy.id_bho_fk
+            """.format(
+                joins_string, where_string
+            )
+
+        if len(lithostratigraphy_params) > 0:
+
+            joins_string = "\n".join(lithostratigraphy_joins) if len(
+                lithostratigraphy_joins) > 0 else ''
+            where_string = (
+                "AND {}".format(" AND ".join(lithostratigraphy_where))
+                if len(lithostratigraphy_where) > 0
+                else ''
+            )
+
+            wr_strt += """
+                INNER JOIN (
+                    SELECT DISTINCT
+                        id_bho_fk
+
+                    FROM
+                        bdms.stratigraphy
+                    
+                    INNER JOIN
+                        bdms.lithostratigraphy
+                    ON
+                        stratigraphy_id = id_sty
+                    {}
+
+                    WHERE
+                        kind_id_cli = 3000
+
+                    {}
+                ) as lithostratigraphy
+                ON 
+                    borehole.id_bho = lithostratigraphy.id_bho_fk
             """.format(
                 joins_string, where_string
             )
@@ -327,8 +364,8 @@ class ListGeojson(Action):
                             ), '{}'::json[]
                         ) AS features
                 ) t
-        """ % (wr_strt, wr), *(layer_params + chronostratigraphy_params + casing_params +
-                   backfill_params + instrument_params + params))
+        """ % (wr_strt, wr), *(layer_params + chronostratigraphy_params + lithostratigraphy_params + 
+                    casing_params + backfill_params + instrument_params + params))
         return {
             "data": self.decode(rec[0])
         }