+++ src/client/src/commons/search/editor/searchEditorState.js
@@ -88,6 +88,9 @@ const initialState = {
     // Chronostratigraphy filter
     chronostratigraphy_id: null,
 
+    // Lithostratigraphy filter
+    lithostratigraphy_id: null,
+
     casing_name: "",
     date_abd_from: null,
     date_abd_to: null,